Refatoração em LspConnection
- Construtor agora proíbe parâmetros nulos, lançando uma exceção
- Tipo de id alterado para short (para ficar igual a Pack)
- Classe StatusChecker convertido para uma função.

+++ test/lsp/LspConnectionTest.java
@@ -30,7 +30,7 @@ public class LspConnectionTest {
 		// alterar closed para true no final.
 		time = 0;
 		closed = false;
-		new LspConnection(1, new LspParams(2, 3), new Actions() {
+		new LspConnection((short) 1, new LspParams(2, 3), new Actions() {
 			public long lastReceiptTime() {
 				return 1 / ++time;
 			}
